requirements.txt
pandas
streamlit
requests
azure-storage-blob